# src/main/java/com/reeceTech/addressBook/entities/AddressBook.java
package com.reeceTech.addressBook.entities;

import jakarta.persistence.Column;
import jakarta.persistence.Entity;
import jakarta.persistence.Id;
import java.io.Serializable;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Entity
@NoArgsConstructor
@AllArgsConstructor
public class AddressBook implements Serializable {
  @Id @Column private String id;

  @Column private String name;

  @Column private String description;
}